+++ src/app/home/home.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { ListaPersonagensComponent } from '../personagens/lista-personagens.component';
+import { ListaPersonagensComponent } from '../personagens/lista-personagens/lista-personagens.component';
 
 @Component({
   selector: 'app-home',
