+++ src/app/personagens/form-personagem/form-personagem.component.ts
@@ -8,9 +8,9 @@ import {
   FormControl,
 } from '@angular/forms';
 import { Router, ActivatedRoute, RouterLink } from '@angular/router';
-import { RickAndMortyServico } from './rick-and-morty.servico';
+import { RickAndMortyServico } from '../rick-and-morty.servico';
 import { MatButtonModule } from '@angular/material/button';
-import { Personagem } from './personagem.model';
+import { Personagem } from '../personagem.model';
 
 @Component({
   selector: 'app-form-personagem',
