+++ src/app/personagens/lista-personagens/lista-personagens.component.scss
