+++ src/app/personagens/lista-personagens/lista-personagens.component.spec.ts
@@ -1,7 +1,7 @@
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
-import { RickAndMortyServico } from './rick-and-morty.servico';
+import { RickAndMortyServico } from '../rick-and-morty.servico';
 import { ListaPersonagensComponent } from './lista-personagens.component';
 import { of } from 'rxjs';
